hadoopcompatibility: Modified hadoopcompatibility stucture in accordance to #777.
