+++ stratosphere-addons/hadoop-compatibility/src/main/java/eu/stratosphere/hadoopcompatibility/mapred/record/example/WordCount.java
@@ -11,32 +11,32 @@
  * specific language governing permissions and limitations under the License.
  **********************************************************************************************************************/
 
-package eu.stratosphere.hadoopcompatibility.example;
+package eu.stratosphere.hadoopcompatibility.mapred.record.example;
 
 import java.io.Serializable;
 import java.util.Iterator;
 import java.util.StringTokenizer;
 
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.TextInputFormat;
-import org.apache.hadoop.mapred.TextOutputFormat;
 
 import eu.stratosphere.api.common.Plan;
 import eu.stratosphere.api.common.Program;
 import eu.stratosphere.api.common.ProgramDescription;
+import eu.stratosphere.api.common.operators.FileDataSink;
 import eu.stratosphere.api.java.record.functions.FunctionAnnotation.ConstantFields;
 import eu.stratosphere.api.java.record.functions.MapFunction;
 import eu.stratosphere.api.java.record.functions.ReduceFunction;
+import eu.stratosphere.api.java.record.io.CsvOutputFormat;
 import eu.stratosphere.api.java.record.operators.MapOperator;
 import eu.stratosphere.api.java.record.operators.ReduceOperator;
 import eu.stratosphere.api.java.record.operators.ReduceOperator.Combinable;
 import eu.stratosphere.client.LocalExecutor;
-import eu.stratosphere.hadoopcompatibility.HadoopDataSink;
-import eu.stratosphere.hadoopcompatibility.HadoopDataSource;
+import eu.stratosphere.hadoopcompatibility.mapred.record.HadoopDataSource;
+import eu.stratosphere.hadoopcompatibility.mapred.record.datatypes.WritableWrapperConverter;
 import eu.stratosphere.types.IntValue;
 import eu.stratosphere.types.Record;
 import eu.stratosphere.types.StringValue;
@@ -46,8 +46,10 @@ import eu.stratosphere.util.Collector;
  * Implements a word count which takes the input file and counts the number of
  * the occurrences of each word in the file.
  */
-@SuppressWarnings("serial")
-public class WordCountWithHadoopOutputFormat implements Program, ProgramDescription {
+public class WordCount implements Program, ProgramDescription {
+	
+	private static final long serialVersionUID = 1L;
+
 
 	/**
 	 * Converts a Record containing one string in to multiple string/integer pairs.
@@ -56,19 +58,19 @@ public class WordCountWithHadoopOutputFormat implements Program, ProgramDescript
 	 */
 	public static class TokenizeLine extends MapFunction implements Serializable {
 		private static final long serialVersionUID = 1L;
-
+		
 		@Override
 		public void map(Record record, Collector<Record> collector) {
 			// get the first field (as type StringValue) from the record
 			String line = record.getField(1, StringValue.class).getValue();
 			// normalize the line
 			line = line.replaceAll("\\W+", " ").toLowerCase();
-
+			
 			// tokenize the line
 			StringTokenizer tokenizer = new StringTokenizer(line);
 			while (tokenizer.hasMoreTokens()) {
 				String word = tokenizer.nextToken();
-
+				
 				// we emit a (word, 1) pair 
 				collector.collect(new Record(new StringValue(word), new IntValue(1)));
 			}
@@ -82,9 +84,9 @@ public class WordCountWithHadoopOutputFormat implements Program, ProgramDescript
 	@Combinable
 	@ConstantFields(0)
 	public static class CountWords extends ReduceFunction implements Serializable {
-
+		
 		private static final long serialVersionUID = 1L;
-
+		
 		@Override
 		public void reduce(Iterator<Record> records, Collector<Record> out) throws Exception {
 			Record element = null;
@@ -98,7 +100,7 @@ public class WordCountWithHadoopOutputFormat implements Program, ProgramDescript
 			element.setField(1, new IntValue(sum));
 			out.collect(element);
 		}
-
+		
 		@Override
 		public void combine(Iterator<Record> records, Collector<Record> out) throws Exception {
 			// the logic is the same as in the reduce function, so simply call the reduce method
@@ -107,30 +109,39 @@ public class WordCountWithHadoopOutputFormat implements Program, ProgramDescript
 	}
 
 
+	@SuppressWarnings({ "rawtypes", "unchecked", "unused" })
 	@Override
 	public Plan getPlan(String... args) {
 		// parse job parameters
 		int numSubTasks   = (args.length > 0 ? Integer.parseInt(args[0]) : 1);
 		String dataInput = (args.length > 1 ? args[1] : "");
 		String output    = (args.length > 2 ? args[2] : "");
-
-		HadoopDataSource<LongWritable, Text> source = new HadoopDataSource<LongWritable, Text>(
-				new TextInputFormat(), new JobConf(), "Input Lines");
+		
+		
+		HadoopDataSource source = new HadoopDataSource(new TextInputFormat(), new JobConf(), "Input Lines");
 		TextInputFormat.addInputPath(source.getJobConf(), new Path(dataInput));
-
-
+		
+		// Example with Wrapper Converter
+		HadoopDataSource<LongWritable,Text> sourceHadoopType = new HadoopDataSource<LongWritable, Text>(
+				new TextInputFormat(), new JobConf(), "Input Lines", new WritableWrapperConverter<LongWritable, Text>());
+		TextInputFormat.addInputPath(source.getJobConf(), new Path(dataInput));
+		
 		MapOperator mapper = MapOperator.builder(new TokenizeLine())
-				.input(source)
-				.name("Tokenize Lines")
-				.build();
+			.input(source)
+			.name("Tokenize Lines")
+			.build();
 		ReduceOperator reducer = ReduceOperator.builder(CountWords.class, StringValue.class, 0)
-				.input(mapper)
-				.name("Count Words")
-				.build();
-		HadoopDataSink<Text, IntWritable> out = new HadoopDataSink<Text, IntWritable>(new TextOutputFormat<Text, IntWritable>(),new JobConf(), "Hadoop TextOutputFormat", reducer, Text.class, IntWritable.class);
-		TextOutputFormat.setOutputPath(out.getJobConf(), new Path(output));
-
-		Plan plan = new Plan(out, "Hadoop OutputFormat Example");
+			.input(mapper)
+			.name("Count Words")
+			.build();
+		FileDataSink out = new FileDataSink(new CsvOutputFormat(), output, reducer, "Word Counts");
+		CsvOutputFormat.configureRecordFormat(out)
+			.recordDelimiter('\n')
+			.fieldDelimiter(' ')
+			.field(StringValue.class, 0)
+			.field(IntValue.class, 1);
+		
+		Plan plan = new Plan(out, "WordCount Example");
 		plan.setDefaultParallelism(numSubTasks);
 		return plan;
 	}
@@ -141,17 +152,17 @@ public class WordCountWithHadoopOutputFormat implements Program, ProgramDescript
 		return "Parameters: [numSubStasks] [input] [output]";
 	}
 
-
+	
 	public static void main(String[] args) throws Exception {
-		WordCountWithHadoopOutputFormat wc = new WordCountWithHadoopOutputFormat();
-
+		WordCount wc = new WordCount();
+		
 		if (args.length < 3) {
 			System.err.println(wc.getDescription());
 			System.exit(1);
 		}
-
+		
 		Plan plan = wc.getPlan(args);
-
+		
 		// This will execute the word-count embedded in a local context. replace this line by the commented
 		// succeeding line to send the job to a local installation or to a cluster for execution
 		LocalExecutor.execute(plan);
